@@ -1,5 +1,5 @@
 import 'package:args/args.dart';
-import 'package:gen_lang/core_18n.dart';
+import 'package:gen_lang_updated/core_18n.dart';
 
 main(List<String> args) async {
   if (_isHelpCommand(args)) {
@@ -25,24 +25,25 @@ I18nOption _generateI18Option(List<String> args) {
   return i18nOption;
 }
 
-ArgParser _generateArgParser(I18nOption i18nOption) {
+ArgParser _generateArgParser(I18nOption? i18nOption) {
   var parser = new ArgParser();
 
-  parser.addOption('source-dir',
-      defaultsTo: 'res/string',
-      callback: (String x) => i18nOption.sourceDir = x,
-      help: 'A source folder contains all string json files');
+  if(i18nOption != null) {
+    parser.addOption('source-dir',
+        defaultsTo: 'res/string',
+        callback: (String? x) => i18nOption.sourceDir = x,
+        help: 'A source folder contains all string json files');
 
-  parser.addOption('output-dir',
-      defaultsTo: 'lib/generated',
-      callback: (String x) => i18nOption.outputDir = x,
-      help: 'A output folder stores all generated files');
+    parser.addOption('output-dir',
+        defaultsTo: 'lib/generated',
+        callback: (String? x) => i18nOption.outputDir = x,
+        help: 'A output folder stores all generated files');
 
-  parser.addOption('template-locale',
-      defaultsTo: 'en',
-      callback: (String x) => i18nOption.templateLocale = x,
-      help:
-          'Use string_{template-locale}.json as a template for default value when a locale does not exist. If string_en does not exist, this script will use the first string json file as a template');
-
+    parser.addOption('template-locale',
+        defaultsTo: 'en',
+        callback: (String? x) => i18nOption.templateLocale = x,
+        help:
+        'Use string_{template-locale}.json as a template for default value when a locale does not exist. If string_en does not exist, this script will use the first string json file as a template');
+  }
   return parser;
 }
