@@ -2,9 +2,9 @@ RegExp ARG_REG_EXP = RegExp(r'\${\w+}');
 const String DEFAULT_PLURAL_ARGS = 'howMany';
 const String DEFAULT_GENDER_ARG = 'targetGender';
 
-List<String> getArgs(Iterable<Match> allMatch, String defaultArg) {
+List<String> getArgs(Iterable<Match> allMatch, String? defaultArg) {
   List<String> args = [];
-  if (null != defaultArg) {
+  if (defaultArg != null) {
     args.add(defaultArg);
   }
 
@@ -18,7 +18,7 @@ List<String> getArgs(Iterable<Match> allMatch, String defaultArg) {
   return args;
 }
 
-String normalizedSpecialCharacters(String message) {
+String? normalizedSpecialCharacters(String? message) {
   if (null != message) {
     String normalizedJson = message.replaceAll(r"\", r"\\");
     return normalizedJson.replaceAll(r'\\"', r'\\\"');
@@ -26,8 +26,8 @@ String normalizedSpecialCharacters(String message) {
   return null;
 }
 
-String normalizedJsonMessage(String message) {
-  if (null != message) {
+String? normalizedJsonMessage(String? message) {
+  if (message != null) {
     return message;
   }
   return null;
@@ -37,15 +37,14 @@ String generateArg(arg) {
   return null != arg ? '"$arg"' : 'null';
 }
 
-String extraArgsFromGender(String male, String female, String other) {
-  List<String> plurals = [male, female, other];
+String extraArgsFromGender(String? male, String? female, String? other) {
+  List<String?> plurals = [male, female, other];
   Iterable<Match> theMostMatch = [];
 
-  for (String plural in plurals) {
+  for (String? plural in plurals) {
     if (null != plural) {
       Iterable<Match> allMatch = ARG_REG_EXP.allMatches(plural);
-      if (null == theMostMatch ||
-          (null != theMostMatch && allMatch.length > theMostMatch.length)) {
+      if (allMatch.length > theMostMatch.length) {
         theMostMatch = allMatch;
       }
     }
@@ -63,16 +62,15 @@ String extraArgsFromGender(String male, String female, String other) {
   return builder.toString();
 }
 
-String extraArgsFromPlural(String zero, String one, String two, String few,
-    String many, String other) {
-  List<String> plurals = [zero, one, two, few, many, other];
+String extraArgsFromPlural(String? zero, String? one, String? two, String? few,
+    String? many, String? other) {
+  List<String?> plurals = [zero, one, two, few, many, other];
   Iterable<Match> theMostMatch = [];
 
-  for (String plural in plurals) {
-    if (null != plural) {
+  for (String? plural in plurals) {
+    if (plural != null) {
       Iterable<Match> allMatch = ARG_REG_EXP.allMatches(plural);
-      if (null == theMostMatch ||
-          (null != theMostMatch && allMatch.length > theMostMatch.length)) {
+      if (allMatch.length > theMostMatch.length) {
         theMostMatch = allMatch;
       }
     }
@@ -90,9 +88,9 @@ String extraArgsFromPlural(String zero, String one, String two, String few,
   return builder.toString();
 }
 
-String extraArgsFromMessage(String message) {
+String extraArgsFromMessage(String? message) {
   StringBuffer builder = StringBuffer();
-  List<String> args = getArgs(ARG_REG_EXP.allMatches(message), null);
+  List<String> args = getArgs(ARG_REG_EXP.allMatches(message ?? ''), null);
 
   for (int i = 0; i < args.length; i++) {
     if (i != 0) {
