@@ -1,4 +1,4 @@
-import 'package:gen_lang/extra_json_message_tool.dart';
+import 'package:gen_lang_updated/extra_json_message_tool.dart';
 
 String generateMessageAllDart(
     String createdMessageLookup, String deferredLibraries, String findExacts) {
@@ -82,16 +82,16 @@ String generateKeyWithValue(String key, String value) {
   return '''\t\t"$key" : $value,''';
 }
 
-String generateSimpleMessage(String message) {
+String generateSimpleMessage(String? message) {
   return '''MessageLookupByLibrary.simpleMessage("${normalizedJsonMessage(message)}")''';
 }
 
-String generateMessageFunction(String args, String message) {
+String generateMessageFunction(String args, String? message) {
   return '''($args) => "${normalizedJsonMessage(message)}"''';
 }
 
-String generatePluralFunction(String args, String zero, String one, String two,
-    String few, String many, String other) {
+String generatePluralFunction(String args, String? zero, String? one, String? two,
+    String? few, String? many, String? other) {
   var zeroArg = generateArg(normalizedJsonMessage(zero));
   var oneArg = generateArg(normalizedJsonMessage(one));
   var twoArg = generateArg(normalizedJsonMessage(two));
@@ -103,7 +103,7 @@ String generatePluralFunction(String args, String zero, String one, String two,
 }
 
 String generateGenderFunction(
-    String args, String male, String female, String other) {
+    String args, String? male, String? female, String? other) {
   var maleArg = generateArg(normalizedJsonMessage(male));
   var femaleArg = generateArg(normalizedJsonMessage(female));
   var otherArg = generateArg(normalizedJsonMessage(other));
@@ -121,7 +121,8 @@ String generateFindExact(String locale) {
         return _\$$locale;''';
 }
 
-bool hasArgsInMessage(String message) {
+bool hasArgsInMessage(String? message) {
+  if(message == null) return false;
   return ARG_REG_EXP.hasMatch(message);
 }
 
